♻️ refactor(metric): introduce standalone metric dispatch API and rename matrix/metric types (#524)
Refactor the metric system around standalone dispatch-based APIs, dedicated metric matrix types, and modular QMatrix infrastructure.

Highlights:
- Rename QuantityMatrix → QMatrix and split implementation into a dedicated package
- Add public det/inv operations for QMatrix
- Introduce coordinax._src.metric with:
  - AbstractMetricField / AbstractDiagonalMetricField
  - AbstractMetricMatrix, DiagonalMetric, DenseMetric
  - metric_matrix and metric_representation dispatch APIs
- Remove metric_matrix methods from metric classes in favour of standalone dispatch
- Add per-geometry register_metric dispatch modules for Euclidean, spherical,
  Minkowski, product, embedded, and custom geometries
- Rename metric classes:
  - EuclideanMetric → FlatMetric
  - HyperSphericalMetric → RoundMetric
  - CartesianProductMetric → ProductMetric
  - InducedMetric → PullbackMetric
- Add scale_factors modules for spherical and Minkowski geometries
- Update AbstractManifold and manifold APIs to use AbstractMetricField
- Export new metric dispatch APIs and matrix types from public modules
- Propagate QMatrix and metric renames throughout charts, vectors,
  transforms, representations, frames, distances, and tests
- Update Astro, CurveFrames, Hypothesis, and Astropy interop packages
  for the new metric and frame APIs
- Refresh specs, guides, tutorials, and README for the metric refactor
- Add typing/configuration fixes and lint/test cleanups
- Simplify numeric literals in tests where dtype semantics permit

Also includes extensive documentation cleanup, doctest normalization,
dispatch registration updates, and compatibility fixes across the codebase.

+++ packages/coordinax.curveframes/src/coordinax/curveframes/_src/base.py
@@ -300,7 +300,7 @@ class AbstractParallelTransportTransform(cxt.AbstractCompositeTransform):
         Q([-0.,  1.,  0.], '')
 
         """
-        R = self._rotation_matrix(tau)
+        R = self._rotation_matrix(tau.astype(float))
         return u.Q(R[0], "")
 
     # ---------------------------------------------------------------
